Use OwnProps in ControlBarContainer
- Removed duplicate properties in Workspace's StateProps
- Used OwnProps in the component as well (see #104)
