+++ src/containers/workspace/ControlBarContainer.ts
@@ -9,12 +9,19 @@ import {
   evalRepl,
   generateLzString
 } from '../../actions/playground'
-import ControlBar, {
-  DispatchProps,
-  OwnProps,
-  StateProps
-} from '../../components/workspace/ControlBar'
+import ControlBar, { DispatchProps, StateProps } from '../../components/workspace/ControlBar'
 import { IState } from '../../reducers/states'
+
+export type OwnProps = {
+  hasChapterSelect?: boolean
+  hasNextButton?: boolean
+  hasPreviousButton?: boolean
+  hasSaveButton?: boolean
+  hasShareButton?: boolean
+  onClickNext?(): any
+  onClickPrevious?(): any
+  onClickSave?(): any
+}
 
 const mapStateToProps: MapStateToProps<StateProps, OwnProps, IState> = (state, props) => ({
   ...props,
@@ -36,4 +43,6 @@ const mapDispatchToProps: MapDispatchToProps<DispatchProps, {}> = (dispatch: Dis
     dispatch
   )
 
-export default connect(mapStateToProps, mapDispatchToProps)(ControlBar)
+export default connect<StateProps, DispatchProps, OwnProps>(mapStateToProps, mapDispatchToProps)(
+  ControlBar
+)
